requirements.txt
pymupdf>=1.26.4
python-dotenv>=1.1.0
groq>=0.37.0
